+++ crates/sncast/tests/e2e/declare_from.rs
@@ -121,7 +121,7 @@ async fn test_contract_already_declared() {
         output,
         indoc! {r"
         Command: declare-from
-        Error: Contract with the same class hash is already declared
+        Error: Contract with class hash 0x0[..] is already declared
         "},
     );
 }